# Rubocop upgrade, Ruby 2.4 minimum and TruffleRuby