--- a/Gemfile
+++ b/Gemfile
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
 git_source(:github) do |repo_name|
   "https://github.com/#{repo_name}.git"
@@ -9,15 +11,15 @@ group :benchmark, :test do
   gem 'benchmark-ips'
   gem 'memory_profiler'
 
-  install_if -> { RUBY_PLATFORM !~ /mingw|mswin|java/ } do
+  install_if -> { RUBY_PLATFORM !~ /mingw|mswin|java/ && RUBY_ENGINE != 'truffleruby' } do
     gem 'stackprof'
   end
 end
 
 group :test do
-  gem 'rubocop', '~> 0.53.0'
+  gem 'rubocop', '~> 0.74.0', require: false
 
-  platform :mri do
-    gem 'liquid-c', github: 'Shopify/liquid-c', ref: '9168659de45d6d576fce30c735f857e597fa26f6'
+  platform :mri, :truffleruby do
+    gem 'liquid-c', github: 'Shopify/liquid-c', ref: '7ba926791ef8411984d0f3e41c6353fd716041c6'
   end
 end