--- a/lib/liquid/block.rb
+++ b/lib/liquid/block.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Liquid
   class Block < Tag
     MAX_DEPTH = 100
@@ -27,16 +29,16 @@ module Liquid
     end
 
     def unknown_tag(tag, _params, _tokens)
-      if tag == 'else'.freeze
-        raise SyntaxError.new(parse_context.locale.t("errors.syntax.unexpected_else".freeze,
-          block_name: block_name))
-      elsif tag.start_with?('end'.freeze)
-        raise SyntaxError.new(parse_context.locale.t("errors.syntax.invalid_delimiter".freeze,
-          tag: tag,
-          block_name: block_name,
-          block_delimiter: block_delimiter))
+      if tag == 'else'
+        raise SyntaxError, parse_context.locale.t('errors.syntax.unexpected_else',
+                                                  block_name: block_name)
+      elsif tag.start_with?('end')
+        raise SyntaxError, parse_context.locale.t('errors.syntax.invalid_delimiter',
+                                                  tag: tag,
+                                                  block_name: block_name,
+                                                  block_delimiter: block_delimiter)
       else
-        raise SyntaxError.new(parse_context.locale.t("errors.syntax.unknown_tag".freeze, tag: tag))
+        raise SyntaxError, parse_context.locale.t('errors.syntax.unknown_tag', tag: tag)
       end
     end
 
@@ -51,18 +53,15 @@ module Liquid
     protected
 
     def parse_body(body, tokens)
-      if parse_context.depth >= MAX_DEPTH
-        raise StackLevelError, "Nesting too deep".freeze
-      end
+      raise StackLevelError, 'Nesting too deep' if parse_context.depth >= MAX_DEPTH
+
       parse_context.depth += 1
       begin
         body.parse(tokens, parse_context) do |end_tag_name, end_tag_params|
           @blank &&= body.blank?
 
           return false if end_tag_name == block_delimiter
-          unless end_tag_name
-            raise SyntaxError.new(parse_context.locale.t("errors.syntax.tag_never_closed".freeze, block_name: block_name))
-          end
+          raise SyntaxError, parse_context.locale.t('errors.syntax.tag_never_closed', block_name: block_name) unless end_tag_name
 
           # this tag is not registered with the system
           # pass it to the current block for special handling or error reporting