--- a/lib/liquid/block_body.rb
+++ b/lib/liquid/block_body.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 module Liquid
   class BlockBody
-    FullToken = /\A#{TagStart}#{WhitespaceControl}?\s*(\w+)\s*(.*?)#{WhitespaceControl}?#{TagEnd}\z/om
-    ContentOfVariable = /\A#{VariableStart}#{WhitespaceControl}?(.*?)#{WhitespaceControl}?#{VariableEnd}\z/om
-    WhitespaceOrNothing = /\A\s*\z/
-    TAGSTART = "{%".freeze
-    VARSTART = "{{".freeze
+    FULL_TOKEN = /\A#{TAG_START}#{WHITESPACE_CONTROL}?\s*(\w+)\s*(.*?)#{WHITESPACE_CONTROL}?#{TAG_END}\z/om.freeze
+    CONTENT_OF_VARIABLE = /\A#{VARIABLE_START}#{WHITESPACE_CONTROL}?(.*?)#{WHITESPACE_CONTROL}?#{VARIABLE_END}\z/om.freeze
+    WHITESPACE_OR_NOTHING = /\A\s*\z/.freeze
+    TAG_START = '{%'
+    VAR_START = '{{'
 
     attr_reader :nodelist
 
@@ -15,36 +17,34 @@ module Liquid
 
     def parse(tokenizer, parse_context)
       parse_context.line_number = tokenizer.line_number
-      while token = tokenizer.shift
+      while (token = tokenizer.shift)
         next if token.empty?
+
         case
-        when token.start_with?(TAGSTART)
+        when token.start_with?(TAG_START)
           whitespace_handler(token, parse_context)
-          unless token =~ FullToken
-            raise_missing_tag_terminator(token, parse_context)
-          end
-          tag_name = $1
-          markup = $2
+          raise_missing_tag_terminator(token, parse_context) unless token =~ FULL_TOKEN
+          tag_name = Regexp.last_match(1)
+          markup = Regexp.last_match(2)
           # fetch the tag from registered blocks
-          unless tag = registered_tags[tag_name]
+          unless (tag = registered_tags[tag_name])
             # end parsing if we reach an unknown tag and let the caller decide
             # determine how to proceed
             return yield tag_name, markup
           end
+
           new_tag = tag.parse(tag_name, markup, tokenizer, parse_context)
           @blank &&= new_tag.blank?
           @nodelist << new_tag
-        when token.start_with?(VARSTART)
+        when token.start_with?(VAR_START)
           whitespace_handler(token, parse_context)
           @nodelist << create_variable(token, parse_context)
           @blank = false
         else
-          if parse_context.trim_whitespace
-            token.lstrip!
-          end
+          token.lstrip! if parse_context.trim_whitespace
           parse_context.trim_whitespace = false
           @nodelist << token
-          @blank &&= !!(token =~ WhitespaceOrNothing)
+          @blank &&= !!(token =~ WHITESPACE_OR_NOTHING)
         end
         parse_context.line_number = tokenizer.line_number
       end
@@ -53,13 +53,11 @@ module Liquid
     end
 
     def whitespace_handler(token, parse_context)
-      if token[2] == WhitespaceControl
+      if token[2] == WHITESPACE_CONTROL
         previous_token = @nodelist.last
-        if previous_token.is_a? String
-          previous_token.rstrip!
-        end
+        previous_token.rstrip! if previous_token.is_a?(String)
       end
-      parse_context.trim_whitespace = (token[-3] == WhitespaceControl)
+      parse_context.trim_whitespace = (token[-3] == WHITESPACE_CONTROL)
     end
 
     def blank?
@@ -67,14 +65,14 @@ module Liquid
     end
 
     def render(context)
-      render_to_output_buffer(context, '')
+      render_to_output_buffer(context, ''.dup)
     end
 
     def render_to_output_buffer(context, output)
       context.resource_limits.render_score += @nodelist.length
 
       idx = 0
-      while node = @nodelist[idx]
+      while (node = @nodelist[idx])
         previous_output_size = output.bytesize
 
         case node
@@ -83,7 +81,7 @@ module Liquid
         when Variable
           render_node(context, output, node)
         when Block
-          render_node(context, node.blank? ? '' : output, node)
+          render_node(context, node.blank? ? ''.dup : output, node)
           break if context.interrupt? # might have happened in a for-block
         when Continue, Break
           # If we get an Interrupt that means the block must stop processing. An
@@ -117,11 +115,12 @@ module Liquid
     def raise_if_resource_limits_reached(context, length)
       context.resource_limits.render_length += length
       return unless context.resource_limits.reached?
-      raise MemoryError.new("Memory limits exceeded".freeze)
+
+      raise MemoryError, 'Memory limits exceeded'
     end
 
     def create_variable(token, parse_context)
-      token.scan(ContentOfVariable) do |content|
+      token.scan(CONTENT_OF_VARIABLE) do |content|
         markup = content.first
         return Variable.new(markup, parse_context)
       end
@@ -129,11 +128,11 @@ module Liquid
     end
 
     def raise_missing_tag_terminator(token, parse_context)
-      raise SyntaxError.new(parse_context.locale.t("errors.syntax.tag_termination".freeze, token: token, tag_end: TagEnd.inspect))
+      raise SyntaxError, parse_context.locale.t('errors.syntax.tag_termination', token: token, tag_end: TAG_END.inspect)
     end
 
     def raise_missing_variable_terminator(token, parse_context)
-      raise SyntaxError.new(parse_context.locale.t("errors.syntax.variable_termination".freeze, token: token, tag_end: VariableEnd.inspect))
+      raise SyntaxError, parse_context.locale.t('errors.syntax.variable_termination', token: token, tag_end: VARIABLE_END.inspect)
     end
 
     def registered_tags