--- a/lib/liquid/extensions.rb
+++ b/lib/liquid/extensions.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'time'
 require 'date'