I'm not sure if this is something that should even be done, but the motivation was originally just to bump the rubocop version to allow `# frozen_string_literal: true` as freezing strings has come up so many times recently across commits and was trying to solve in another pull request.

The most important breaking changes are the shift away from class variables `@@var` and the shift of all constants to be `SCREAMING_SNAKE_CASE`

This is now following https://shopify.github.io/ruby-style-guide/ with the single exception that the following is defined

```
Metrics/LineLength:
  Enabled: false
```

This also bumps the minimum Ruby version up to `2.3` based on Rubocop settings

@Shopify/liquid @Shopify/guardians-of-the-liquid