@shopmike 2019-08-26
Seems to be a slight performance boost, maybe from the universal freeze, although could just be margin of error

**BEFORE**
```
              parse:     34.582  (± 2.9%) i/s -    348.000  in  10.068965s
             render:     91.771  (± 3.3%) i/s -    920.000  in  10.034738s
     parse & render:     23.452  (± 4.3%) i/s -    236.000  in  10.072722s


              parse:     34.894  (± 2.9%) i/s -    351.000  in  10.065514s
             render:     91.119  (± 2.2%) i/s -    918.000  in  10.081380s
     parse & render:     23.849  (± 4.2%) i/s -    240.000  in  10.070867s
```
**AFTER**
```
              parse:     35.909  (± 2.8%) i/s -    360.000  in  10.032490s
             render:     93.636  (± 2.1%) i/s -    936.000  in  10.000736s
     parse & render:     24.535  (± 4.1%) i/s -    246.000  in  10.035064s

              parse:     35.480  (± 2.8%) i/s -    357.000  in  10.070159s
             render:     92.217  (± 3.3%) i/s -    927.000  in  10.060456s
     parse & render:     24.815  (± 0.0%) i/s -    250.000  in  10.077457s
```

Also added support so `TruffleRuby` now passes on CI (Disabled Rubocop when target isn't MRI). Although it seems like the C extension isn't actually loading as its not outputting any strings that its enabled.