@fw42 2019-08-26
Should we consider this to be a non-backwards-compatible change (because of all the constants that got renamed, which third parties like Jekyll might depend on)?

> Also added support so TruffleRuby now passes on CI (Disabled Rubocop when target isn't MRI). Although it seems like the C extension isn't actually loading as its not outputting any strings that its enabled.

If Truffle is passing CI (I assume this means that no tests are failing?), then why is it listed under `allow_failures`?