@shopmike 2019-08-26
Yes in its current form this is a breaking change because of the constants, that will impact almost all custom tags as they generally use `VariableSignature` which is now `VARIABLE_SIGNATURE`. This is a decision that needs to be made as these can be rolled back and an exception added to the rubocop config. I just went for full support to see what it would look like.

I put Truffle under `allow_failures` as it was failing and wasn't sure if it would pass. I'm also not sure if you can commit to officially supporting Truffle as I'm not sure its even classed as stable anyway.