@fw42 2019-08-26
Not sure we want to officially "support" Truffle at this point yet but it would be nice to not accidentally break it again now that it's already passing. I'd say lets remove it from the `allow_failures` list for now so it's very visible when we introduce regressions? @pushrax? FYI @chrisseaton