@tobi 2019-08-26
agreed flo.

I think we can rename what we want, lets offer am optional require
liquid/legacy which does renames to old constant names. This also allows us
to deprecate some stuff in the future.

- tobi
CEO Shopify


On Mon, Aug 26, 2019 at 9:57 AM Florian Weingarten <notifications@github.com>
wrote:

> Not sure we want to officially "support" Truffle at this point yet but it
> would be nice to not accidentally break it again now that it's already
> passing. I'd say lets remove it from the allow_failures list for now so
> it's very visible when we introduce regressions? @pushrax
> <https://github.com/pushrax>? FYI @chrisseaton
> <https://github.com/chrisseaton>
>
> —
> You are receiving this because you are on a team that was mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Shopify/liquid/pull/1130?email_source=notifications&email_token=AAAACW6GXIISWAVWUQVXMMTQGPONRA5CNFSM4IPOOS32YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD5EOJ7A#issuecomment-524870908>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAAACW3UAGK7L3EXZRVBQ6DQGPONRANCNFSM4IPOOS3Q>
> .
>