@shopmike 2019-08-26
Need to play around a little more with the TruffleRuby CI. It isn't consistent when it comes to the liquid-c library. 

For some reason it's not picking up the environmental variable on the CI, but it is on my local now that I've installed it. 

The ruby side runs fine and all tests pass, the C version just seems to be failing due to `parse_range` locally but the CI isn't even thinking it needs to load it, must be caching or something silly. 

I'll play around a little more tomorrow when I'm back online.