@shopmike 2019-08-26
So I got one step closer. For some reason the CI environment didn't like 
```
ENV['LIQUID-C'] = '1'
```
Dropping the dash to an underscore resolved the problem
```
ENV['LIQUID_C'] = '1'
```