@shopmike 2019-08-26
So a modified version of `liquid-c` is now passing. `rb_responds_to` seems to have a few quirks in TruffleRuby, which no doubt will be fixed in future. I've actually just been able to remove the line that was causing the error as the code just falls back to letting Ruby handle range lookups.

There is also one test that intermittently fails `test_does_not_add_drop_methods_to_symbol_table` as it exposes methods on drops.

<img width="648" alt="Screen Shot 2019-08-27 at 3 07 17 am" src="https://user-images.githubusercontent.com/53470248/63708662-63f2d480-c878-11e9-84a1-a88d3e980557.png">