@ashmaroli 2019-08-26
@shopmike RuboCop does allow aligning equal signs and the *"hash rockets"* `=>` to improve readability. However, they're not enabled by default.
```ruby
# bad (IMO)
{
  'foo' => 'bar'
  'alpha => 'beta'
  'marvel' => 'comics'
}

FOO = 'lorem ipsum'
ALPHA = 'the quick brown fox'
MARVEL = 'Good luck, Tiger!'
```
```ruby
# good (IMO)
{
  'foo'    => 'bar'
  'alpha   => 'beta'
  'marvel' => 'comics'
}

FOO    = 'lorem ipsum'
ALPHA  = 'the quick brown fox'
MARVEL = 'Good luck, Tiger!'
```