@shopmike 2019-08-26
@ashmaroli I will leave it up to everyone to decide I'm far from a Rubyist to make that call, I was just sticking to defaults to do the initial clean up. The issue I seemed to have come across and why the commit was out is it seems like the Module level doesn't seem to have the same forced rules as what Class level does. Hopefully someone has the same issue and resolves it in Rubocop.