@shopmike 2019-08-26
Actually just played around, so if 2 lines have = in same position it doesn't raise any issues. As they were previously already partially formatted it didn't think anything was wrong even though its currently set to a single space.