@ashmaroli 2019-08-26
> the Module level doesn't seem to have the same forced rules as what Class level does.

Didn't quite get the draft here. Please clarify a bit more..