@shopmike 2019-08-26
I used `rubocop --auto-correct` for most of the spacing fixes. I thought it played up because those constants weren't in a class they were on the module. But it had nothing to do with that, it was because they were already aligned so it didn't change them to a single space like everything else.