@shopmike 2019-08-27
I've also included the following PR and Issues in this which this pull request supersedes or includes similar.

https://github.com/Shopify/liquid/pull/1096 - Credit @davidcornu 
https://github.com/Shopify/liquid/pull/991 - Credit @dylanahsmith 
https://github.com/Shopify/liquid/pull/990 - Credit @ashmaroli 
https://github.com/Shopify/liquid/pull/751 - Credit @nijikon