@shopmike 2019-08-27
This now bumps minimum ruby version to `2.4` which was mentioned by @davidcornu 
here https://github.com/Shopify/liquid/pull/1096#issue-272899122