@shopmike 2019-08-27
@tobi I've included the legacy support as suggested and this can be enabled by using

```
require 'liquid'
require 'liquid/legacy'
```