@fw42 2019-08-27
Can you clean up the commit history before merging?